// src/api/models/upk-question.ts
/* tslint:disable */
/* eslint-disable */
/**
 * General Research Full Service Brokerage API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { UpkQuestionType } from './upk-question-type';
 /**
 * 
 *
 * @export
 * @interface UpkQuestion
 */
export interface UpkQuestion {

    /**
     * @type {any}
     * @memberof UpkQuestion
     */
    questionId?: any;

    /**
     * what marketplace question this question links to
     *
     * @type {any}
     * @memberof UpkQuestion
     */
    extQuestionId?: any;

    /**
     * @type {UpkQuestionType}
     * @memberof UpkQuestion
     */
    questionType: UpkQuestionType;

    /**
     * @type {any}
     * @memberof UpkQuestion
     */
    countryIso: any;

    /**
     * @type {any}
     * @memberof UpkQuestion
     */
    languageIso: any;

    /**
     * The text shown to respondents
     *
     * @type {any}
     * @memberof UpkQuestion
     */
    questionText: any;

    /**
     * @type {any}
     * @memberof UpkQuestion
     */
    choices?: any;

    /**
     * @type {any}
     * @memberof UpkQuestion
     */
    selector?: any;

    /**
     * @type {any}
     * @memberof UpkQuestion
     */
    configuration?: any;

    /**
     * @type {any}
     * @memberof UpkQuestion
     */
    validation?: any;

    /**
     * @type {any}
     * @memberof UpkQuestion
     */
    importance?: any;
}